backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.1.13
langchain-core==0.1.30
langchain-google-genai==0.0.12
langgraph==0.0.26
websockets==12.0
python-multipart==0.0.6
pymongo==4.6.1